requirements.txt
absl-py==2.1.0
accelerate==1.5.1
annotated-types==0.7.0
array_record==0.7.1
astunparse==1.6.3
attrs==25.3.0
black==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
clarabel==0.10.0
click==8.1.8
cmeel==0.57.1
cmeel-assimp==5.4.3.1
cmeel-boost==1.83.0
cmeel-console-bridge==1.0.2.2
cmeel-octomap==1.10.0
cmeel-qhull==8.0.2.1
cmeel-tinyxml==2.6.2.3
cmeel-urdfdom==3.1.1.1
cmeel-zlib==1.3.1
contourpy==1.3.1
cvxpy==1.6.3
cycler==0.12.1
debugpy==1.8.13
decord==0.6.0
deepspeed==0.16.4
dm-tree==0.1.9
docker-pycreds==0.4.0
docstring_parser==0.16
eigenpy==3.5.1
einops==0.8.1
etils==1.12.2
filelock==3.17.0
flash-attn @ https://github.com/Dao-AILab/flash-attention/releases/download/v2.7.0.post2/flash_attn-2.7.0.post2+cu11torch2.5cxx11abiFALSE-cp312-cp312-linux_x86_64.whl
flatbuffers==25.2.10
fonttools==4.56.0
freetype-py==2.5.1
fsspec==2025.3.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
googleapis-common-protos==1.69.1
grpcio==1.71.0
h5py==3.13.0
hjson==3.1.0
hpp-fcl==2.4.4
huggingface-hub==0.29.3
idna==3.10
imageio==2.37.0
immutabledict==4.2.1
importlib_resources==6.5.2
Jinja2==3.1.6
keras==3.9.0
kiwisolver==1.4.8
libclang==18.1.1
lxml==5.3.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
msgpack==1.1.0
mypy-extensions==1.0.0
namex==0.0.8
networkx==3.4.2
ninja==1.11.1.3
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu11==9.1.0.70
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.570.86
nvidia-nccl-cu11==2.21.5
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.4.127
opt_einsum==3.4.0
optree==0.14.1
osqp==0.6.7.post3
packaging==24.2
pathspec==0.12.1
pillow==11.0.0
pin==2.7.0
platformdirs==4.3.6
pooch==1.8.2
promise==2.3
protobuf==4.25.3
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==19.0.1
pycollada==0.9
pydantic==2.10.6
pydantic_core==2.27.2
pyglet==2.1.3
Pygments==2.19.1
PyOpenGL==3.1.9
pyparsing==3.2.1
pyribbit==0.1.46
python-dateutil==2.9.0.post0
pyvista==0.44.2
PyYAML==6.0.2
pyzmq==26.3.0
qdldl==0.1.7.post5
qpsolvers==4.5.0
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.5.3
scipy==1.15.2
scooby==0.10.0
scs==3.2.7.post2
sentencepiece==0.2.0
sentry-sdk==2.22.0
setproctitle==1.3.5
setuptools==76.1.0
simple-parsing==0.1.7
six==1.17.0
smmap==5.0.2
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-datasets==4.9.8
tensorflow-metadata==1.16.1
termcolor==2.5.0
tf_keras==2.19.0
timm==1.0.15
tokenizers==0.21.1
toml==0.10.2
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.47.0
transforms3d==0.4.2
trimesh==4.6.4
triton==3.1.0
typing_extensions==4.12.2
urchin==0.0.29
urllib3==2.3.0
vtk==9.3.1
wandb==0.19.8
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
zipp==3.21.0
